AC5 - don't show pagination buttons when an error occurs + refactoring
